tehtävä 5.7

+++ src/components/Blog.js
@@ -1,4 +1,7 @@
+import { useState } from 'react'
+
 const Blog = ({ blog, handleLike, user, handleDelete }) => {
+  const [visible, setVisible] = useState(false)
 
   /*console.log(`käyttäjä on ${user.name} ja id ${user}`)*/
   console.log(user)
@@ -6,11 +9,33 @@ const Blog = ({ blog, handleLike, user, handleDelete }) => {
   /*window.confirm(`are you sure you wsant delete blog: ${blog.title}?`
   */
 
-  return(
-    <div>
-      {blog.title} {blog.author} {blog.likes}
-      <button id='like-button' onClick={() => handleLike(blog)}>like</button>
+  const toggleVisibility = () => {
+    setVisible(!visible)
+  }
+
+  const blogStyle = {
+    paddingTop: 10,
+    paddingLeft: 2,
+    border: 'solid',
+    borderWidth: 1,
+    marginBottom: 5
+  }
+
+  if(!visible) return(
+    <div style={blogStyle}>
+      <div>
+        {blog.title} {blog.author}
+      </div>
+      <button onClick={toggleVisibility}>view</button>
+    </div>
+  )
+  return (
+    <div style={blogStyle}>
+      <div>{blog.title} {blog.author}</div>
+      <div>{blog.url}</div>
+      <div>likes: {blog.likes} <button id='like-button' onClick={() => handleLike(blog)}>like</button></div>
       {user !== null && blog.user.id === user.id ? <button id='delete-button' onClick={() => handleDelete(blog)}>delete</button>:<></>}
+      <button onClick={toggleVisibility}>hide</button>
     </div>
   )
 }
